addd .

+++ demo/src/main/java/com/neo4j_ecom/demo/model/entity/ProductReview.java
@@ -4,11 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
 import org.springframework.data.neo4j.core.schema.GeneratedValue;
 import org.springframework.data.neo4j.core.schema.Id;
 import org.springframework.data.neo4j.core.schema.Node;
-import org.springframework.data.neo4j.core.schema.Relationship;
+import org.springframework.data.neo4j.core.schema.Property;
 import org.springframework.data.neo4j.core.support.UUIDStringGenerator;
 
 @AllArgsConstructor
@@ -22,5 +21,5 @@ public class ProductReview {
     @GeneratedValue(UUIDStringGenerator.class)
     private String id;
     private String content;
-    private int rating;
+    private float rating;
 }
